// joycribe-desktop/react-app/src/types.ts
export enum Sentiment {
    Positive = 'Positive',
    Negative = 'Negative',
    Neutral = 'Neutral',
    Mixed = 'Mixed'
}

export enum AnalysisPersona {
    General = 'General Analyst',
    Sales = 'Sales Coach',
    Support = 'Technical Support Supervisor',
    CustomerService = 'Customer Service',
    Personal = 'Personal Chat',
}

export interface SalesCoachingInsights {
    overallRating: number;
    ratingJustification: string;
    strengths: string[];
    improvementAreas: string[];
    talkToListenRatio: string;
    objectionHandling: {
        objection: string;
        suggestedResponse: string;
    }[];
}


export interface SupportInsights {
    problemDescription: string;
    solutionSteps: string[];
    technicalTerms: string[];
    escalationNeeded: string;
    nextSteps: string[];
}


export interface Analysis {
    summary: string;
    sentiment: Sentiment;
    keywords: string[];
    topics: string[];
    customEntities: { entity: string; occurrences: string[] }[];
    salesCoachingInsights?: SalesCoachingInsights;
    supportInsights?: SupportInsights;
    suggestedTags?: string[];
    diarizedTranscript: string;
}